requirements.txt
bleak==0.22.3
dbus-fast==2.24.3
typing_extensions==4.12.2
paho-mqtt==2.1.0
